+++ db/migrate/201408111702_add_association_columns.rb
@@ -1,10 +1,15 @@
 class AddAssociationColumns < ActiveRecord::Migration
 
   def change
-    # add_column :sports, :teams, :has_many
-    add_column :teams, :sport, :belongs_to
-    # add_column :teams, :players, :has_many
-    # add_column :teams, :captain, :has_one
+    change_table "sports", force: true do |t|
+      t.references :teams
+    end
+
+    change_table "teams", force: true do |t|
+      t.references :sport
+      t.references :players
+      t.references :captain
+    end
   end
 
 end