+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 201408111339) do
+ActiveRecord::Schema.define(version: 201408111702) do
 
   create_table "sports", force: true do |t|
     t.string  "name"
@@ -19,10 +19,14 @@ ActiveRecord::Schema.define(version: 201408111339) do
     t.integer "maximum_team_size",       default: 100
     t.integer "minimum_number_of_teams", default: 0
     t.integer "maximum_number_of_teams", default: 100
+    t.integer "teams_id"
   end
 
   create_table "teams", force: true do |t|
-    t.string "name"
+    t.string  "name"
+    t.integer "sport_id"
+    t.integer "players_id"
+    t.integer "captain_id"
   end
 
 end